NLP_Project/requirements.txt
farm-haystack==1.25.2
transformers==4.37.2
sentence-transformers==2.2.2
torch==2.5.0
numpy==1.24.3
huggingface_hub==0.20.3